data to outer files + convert project file to jsx
