+++ src/components/burger-constructor/burger-constructor-list/cunstructor-ingredients-list/constructor-ingredients-list.jsx
@@ -2,6 +2,7 @@ import React from 'react';
 import { ConstructorElement, DragIcon } from "@ya.praktikum/react-developer-burger-ui-components";
 import style from './constructor-ingredients-list.module.css'
 import PropTypes from 'prop-types';
+import { selectedIngredientsType } from "../../../../utils/types";
 
 function ConstructorIngredientsList( { selectedIngredients } ) {
     return (
@@ -24,13 +25,6 @@ function ConstructorIngredientsList( { selectedIngredients } ) {
     );
 }
 
-ConstructorIngredientsList.propTypes = {
-    selectedIngredients: PropTypes.arrayOf(PropTypes.shape({
-        _id: PropTypes.string,
-        name: PropTypes.string,
-        price: PropTypes.number,
-        image: PropTypes.string,
-    }))
-}
+ConstructorIngredientsList.propTypes = selectedIngredientsType
 
 export default ConstructorIngredientsList;
