+++ src/components/burger-ingredients/burger-ingredients-category/burger-ingredients-category.jsx
@@ -2,6 +2,7 @@ import React from 'react';
 import style from './burger-ingredients-category.module.css'
 import BurgerIngredient from "../burger-ingredient/burger-ingredient";
 import PropTypes from "prop-types";
+import { ingredientsType } from "../../../utils/types";
 
 function BurgerIngredientsCategory( { title, ingredients } ) {
 
@@ -17,7 +18,9 @@ function BurgerIngredientsCategory( { title, ingredients } ) {
                             name={ ingredient.name }
                             price={ ingredient.price }
                             img={ ingredient.image }
-                            count={ 1 }/>
+                            count={ 1 }
+                            alt={ ingredient.name }
+                        />
                     ) )
                 }
             </div>
@@ -26,20 +29,7 @@ function BurgerIngredientsCategory( { title, ingredients } ) {
 }
 
 BurgerIngredientsCategory.propTypes = {
-    ingredients: PropTypes.arrayOf( PropTypes.shape( {
-        _id: PropTypes.string,
-        name: PropTypes.string,
-        type: PropTypes.string,
-        proteins: PropTypes.number,
-        fat: PropTypes.number,
-        carbohydrates: PropTypes.number,
-        calories: PropTypes.number,
-        price: PropTypes.number,
-        image: PropTypes.string,
-        image_mobile: PropTypes.string,
-        image_large: PropTypes.string,
-        __v: PropTypes.number,
-    } ) ).isRequired,
+    ...ingredientsType,
     title: PropTypes.string.isRequired,
 }
