+++ src/utils/data.js
@@ -210,3 +210,48 @@ export const data = [
         "__v":0
     }
 ]
+
+export const selectedIngredients = [
+    {
+        "_id": "60666c42cc7b410027a1a9b1",
+        "name": "Краторная булка N-200i",
+        "price": 1255,
+        "image": "https://code.s3.yandex.net/react/code/bun-02.png",
+    },
+    {
+        "_id": "60666c42cc7b410027a1a9b9",
+        "name": "Соус традиционный галактический",
+        "price": 15,
+        "image": "https://code.s3.yandex.net/react/code/sauce-03.png",
+    },
+    {
+        "_id": "60666c42cc7b410027a1a9b4",
+        "name": "Мясо бессмертных моллюсков Protostomia",
+        "price": 1337,
+        "image": "https://code.s3.yandex.net/react/code/meat-02.png",
+    },
+    {
+        "_id": "60666c42cc7b410027a1a9bc",
+        "name": "Плоды Фалленианского дерева",
+        "price": 874,
+        "image": "https://code.s3.yandex.net/react/code/sp_1.png",
+    },
+    {
+        "_id": "60666c42cc7b410027a1a9bb",
+        "name": "Хрустящие минеральные кольца",
+        "price": 300,
+        "image": "https://code.s3.yandex.net/react/code/mineral_rings.png",
+    },
+    {
+        "_id": "60666c42cc7b410027a1a9bb",
+        "name": "Хрустящие минеральные кольца",
+        "price": 300,
+        "image": "https://code.s3.yandex.net/react/code/mineral_rings.png",
+    },
+    {
+        "_id": "60666c42cc7b410027a1a9b1",
+        "name": "Краторная булка N-200i",
+        "price": 1255,
+        "image": "https://code.s3.yandex.net/react/code/bun-02.png",
+    },
+]
