+++ src/utils/types.js
@@ -1,0 +1,27 @@
+import PropTypes from "prop-types";
+
+export const ingredientsType = {
+    ingredients: PropTypes.arrayOf( PropTypes.shape( {
+        _id: PropTypes.string,
+        name: PropTypes.string,
+        type: PropTypes.string,
+        proteins: PropTypes.number,
+        fat: PropTypes.number,
+        carbohydrates: PropTypes.number,
+        calories: PropTypes.number,
+        price: PropTypes.number,
+        image: PropTypes.string,
+        image_mobile: PropTypes.string,
+        image_large: PropTypes.string,
+        __v: PropTypes.number,
+    } ) ).isRequired,
+}
+
+export const selectedIngredientsType = {
+    selectedIngredients: PropTypes.arrayOf(PropTypes.shape({
+        _id: PropTypes.string.isRequired,
+        name: PropTypes.string.isRequired,
+        price: PropTypes.number.isRequired,
+        image: PropTypes.string,
+    }))
+}
